Remove Python 2.7, add Python 3.9

+++ ci/build_linux.sh
@@ -4,6 +4,14 @@ set -e -x
 export DSS_PYTHON_MANYLINUX=1
 cd /io/dss_python
 for PYBIN in /opt/python/*/bin; do
+    if [[ $PYBIN == *"cp27"* ]]
+    then
+        continue
+    fi
+    if [[ $PYBIN == *"cp34"* ]]
+    then
+        continue
+    fi
     rm -rf /io/dss_python/.eggs/
     rm -rf /io/dss_python/build
     "${PYBIN}/python" setup.py --quiet bdist_wheel
